Fixed missing libraries (wrong name). Hack to masq out GDK warnings from beeing reported.
git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@5858 d4735bf7-8e1f-0410-bce4-ef584eb11c25

+++ base/gpe-su/gpe-su.c
@@ -112,7 +112,8 @@ do_su (void)
     }
   *rp = 0;
 
-  if (strncmp (result, "Password:", 9))
+  if ((strncmp (result, "Password:", 9))
+	  && (!strstr(result, "Gdk-WARNING")))
     {
       gpe_error_box (result);
       kill (pid, 15);
